重构 MemoryAllocator 类及相关代码
将 MemoryAllocator 类中的多个方法的参数类型从 int 改为 uint，并重构了 AllocAnsi 和 AllocUni 方法，使用 Utils 类中的新方法计算字节数。在 DxcCompiler 类中添加了 System.Runtime.CompilerServices 的引用，并在 Compile 方法中使用 MemoryAllocator 分配字符串。在 VKFrameBuffer、VKGraphicsPipeline 和 VKResourceLayout 类中相应地更新了 MemoryAllocator 的调用。

+++ src/ZenithEngine.Vulkan/VKFrameBuffer.cs
@@ -19,7 +19,7 @@ internal unsafe class VKFrameBuffer : FrameBuffer
 
         TextureSampleCount sampleCount = TextureSampleCount.Count1;
 
-        RenderingAttachmentInfo* colorAttachmentInfos = MemoryAllocator.Alloc<RenderingAttachmentInfo>(ColorTargets.Length);
+        RenderingAttachmentInfo* colorAttachmentInfos = MemoryAllocator.Alloc<RenderingAttachmentInfo>((uint)ColorTargets.Length);
         RenderingAttachmentInfo* depthStencilAttachmentInfo = null;
 
         PixelFormat[] colorFormats = new PixelFormat[ColorTargets.Length];
