+++ src/ZenithEngine.Vulkan/VKResourceLayout.cs
@@ -12,7 +12,7 @@ internal unsafe class VKResourceLayout : ResourceLayout
     public VKResourceLayout(GraphicsContext context,
                             ref readonly ResourceLayoutDesc desc) : base(context, in desc)
     {
-        DescriptorSetLayoutBinding* bindings = MemoryAllocator.Alloc<DescriptorSetLayoutBinding>(desc.Elements.Length);
+        DescriptorSetLayoutBinding* bindings = MemoryAllocator.Alloc<DescriptorSetLayoutBinding>((uint)desc.Elements.Length);
 
         uint uniformBufferCount = 0;
         uint storageBufferCount = 0;
